refine utils; formalize utils tests
