+++ reference_data/basis_sets/default/sto-3g.nw
@@ -1,6 +1,6 @@
 #----------------------------------------------------------------------
 # Basis Set Exchange
-# Version v0.9
+# Version v0.9.1
 # https://www.basissetexchange.org
 #----------------------------------------------------------------------
 #   Basis set: STO-3G
