+++ tests/python/utils_tests/test_utils.py
@@ -1,0 +1,35 @@
+#
+# Copyright 2023 NWChemEx-Project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import os
+import sys
+import unittest
+
+# Add utils into the path
+my_dir = os.path.dirname(os.path.realpath(__file__))
+utils_path = os.path.join("..", "..", "..", "utils")
+utils_dir = os.path.abspath(os.path.join(my_dir, utils_path))
+sys.path.append(utils_dir)
+
+if __name__ == '__main__':
+
+    my_dir = os.path.dirname(os.path.realpath(__file__))
+
+    loader = unittest.TestLoader()
+    tests = loader.discover(my_dir)
+    testrunner = unittest.runner.TextTestRunner()
+    ret = not testrunner.run(tests).wasSuccessful()
+    sys.exit(ret)
