+++ utils/data_management/generate_basis.py
@@ -54,8 +54,8 @@ import os
 import re
 import sys
 
-from generate_atomicinfo import AtomicData, parse_symbols
-import helper_fxns as helpers
+from data_management.generate_atomicinfo import AtomicData, parse_symbols
+import data_management.helper_fxns as helpers
 
 
 class Shell:
